Code cleanup and refactoring. * Broke out ramrod/__init__.py into ramrod, ramrod.errors,   ramrod.base and ramrod.version. * Broke out ramrod.stix into ramrod.stix, ramrod.stix.common   and ramrod.stix.base. * Broke out ramrod.cybox into ramrod.cybox, ramrod.cybox.common,   and ramrod.cybox.base. * Moved some common, duplicative functions into ramrod.utils. * General bug fixes and cleanup
