+++ docs/api/errors.rst
@@ -1,0 +1,7 @@
+:mod:`ramrod.errors` Module
+===========================
+
+.. automodule:: ramrod.errors
+    :members:
+    :undoc-members:
+    :show-inheritance:
