+++ ramrod/cybox/__init__.py
@@ -1,104 +1,12 @@
 # Copyright (c) 2015, The MITRE Corporation. All rights reserved.
 # See LICENSE.txt for complete terms.
 
+# internal
+import ramrod
 import ramrod.utils as utils
-import ramrod.errors as errors
-from ramrod import (_BaseUpdater, _Vocab, UpdateResults)
 
-TAG_CYBOX_MAJOR  = "cybox_major_version"
-TAG_CYBOX_MINOR  = "cybox_minor_version"
-TAG_CYBOX_UPDATE = "cybox_update_version"
-
-CYBOX_VERSIONS = ('2.0', '2.0.1', '2.1')
-
-
-class _CyboxUpdater(_BaseUpdater):
-    """Base class for CybOX updating code. Sets default values for
-    CybOX-specific xpaths and namespaces.
-
-    """
-    DEFAULT_VOCAB_NAMESPACE = 'http://cybox.mitre.org/default_vocabularies-2'
-    XPATH_VERSIONED_NODES = "//cybox:Observables"
-    XPATH_ROOT_NODES = "//cybox:Observables"
-    XPATH_OBJECT_PROPS = "//cybox:Object/cybox:Properties"
-
-    def __init__(self):
-        super(_CyboxUpdater, self).__init__()
-
-    @classmethod
-    def get_version(cls, observables):
-        """Returns the version of the `observables` ``Observables`` node.
-
-        Returns:
-            A dotted-decimal a version string from the ``cybox_major``,
-            ``cybox_minor`` and ``cybox_update`` attribute values.
-
-        Raises:
-            UnknownVersionError: If `observables` does not contain any of the
-                following attributes:
-
-                * ``cybox_major_version``
-                * ``cybox_minor_version``
-                * ``cybox_update_version``
-        """
-        cybox_major  = observables.attrib.get(TAG_CYBOX_MAJOR)
-        cybox_minor  = observables.attrib.get(TAG_CYBOX_MINOR)
-        cybox_update = observables.attrib.get(TAG_CYBOX_UPDATE)
-
-        if not (cybox_major and cybox_minor):
-            error = "CybOX document contains no version information."
-            raise errors.UnknownVersionError(error)
-
-        if cybox_update:
-            version = "%s.%s.%s" % (cybox_major, cybox_minor, cybox_update)
-        else:
-            version = "%s.%s" % (cybox_major, cybox_minor)
-
-        return version
-
-
-    def _check_version(self, root):
-        """Checks the versions of the Observables instances found in the
-        `root` document. This overrides the ``_BaseUpdater._check_version()``
-        method.
-
-        Note:
-            The ``version`` attribute of `root` is compared against the
-            ``VERSION`` class-level attribute.
-
-        Args:
-            root: The root node for the document.
-
-        Raises:
-            UnknownVersionError: If `root` does not contain a ``version``
-                attribute.
-            InvalidVersionError: If the ``version`` attribute value for `root`
-                does not match the value of ``VERSION``.
-
-        """
-        roots = self._get_root_nodes(root)
-        expected = self.VERSION
-
-        for node in roots:
-            found = self.get_version(node)
-
-            if utils.is_version_equal(expected, found):
-                continue
-
-            error = "Document version '{0}' does not match the expected version '{1}'."
-            error = error.format(found, expected)
-            raise errors.InvalidVersionError(
-                message=error,
-                node=node,
-                expected=expected,
-                found=found
-            )
-
-
-class _CyboxVocab(_Vocab):
-    VOCAB_NAMESPACE = _CyboxUpdater.DEFAULT_VOCAB_NAMESPACE
-
-
+# relative
+from . import base, common
 from .cybox_2_0 import Cybox_2_0_Updater
 from .cybox_2_0_1 import Cybox_2_0_1_Updater
 
@@ -106,6 +14,15 @@ CYBOX_UPDATERS = {
     '2.0': Cybox_2_0_Updater,
     '2.0.1': Cybox_2_0_1_Updater
 }
+
+CYBOX_VERSIONS = common.CYBOX_VERSIONS
+
+
+def get_version(doc):
+    """Returns the version number for input CybOX document."""
+    root = utils.get_etree_root(doc)
+    return base.BaseCyboxUpdater.get_version(root)
+
 
 def update(doc, from_=None, to_=None, options=None, force=False):
     """Updates a CybOX document to align with a given version of the CybOX
@@ -129,27 +46,28 @@ def update(doc, from_=None, to_=None, options=None, force=False):
         An instance of ``ramrod.UpdateResults``.
 
     Raises:
-        ramrod.UpdateError: If any of the following conditions are encountered:
+        .UpdateError: If any of the following conditions are encountered:
 
             * The `from_` or `to_` versions are invalid.
             * An untranslatable field is encountered and `force` is ``False``.
             * A non-unique ID is encountered and `force` is ``False``.
-        ramrod.InvalidVersionError: If the source document version and the
+        .InvalidVersionError: If the source document version and the
             `from_` value do not match and `force` is ``False``.
-        ramrod.UnknownVersionError: If the source document does not contain
+        .UnknownVersionError: If the source document does not contain
             version information and `force` is ``False``.
 
     """
     root = utils.get_etree_root(doc)
-    from_ = from_ or _CyboxUpdater.get_version(root)
-    to_ = to_ or CYBOX_VERSIONS[-1]  # The latest version if not specified
+    versions = common.CYBOX_VERSIONS
+    from_ = from_ or base.BaseCyboxUpdater.get_version(root)
+    to_ = to_ or versions[-1]  # The latest version if not specified
 
-    utils.validate_versions(from_, to_, CYBOX_VERSIONS)
+    utils.validate_versions(from_, to_, versions)
 
     removed, remapped = [], {}
-    idx = CYBOX_VERSIONS.index
+    idx = versions.index
 
-    for version in CYBOX_VERSIONS[idx(from_):idx(to_)]:
+    for version in versions[idx(from_):idx(to_)]:
         updater   = CYBOX_UPDATERS[version]
         results   = updater().update(root, options=options, force=force)
         root      = results.document.as_element()
@@ -157,7 +75,7 @@ def update(doc, from_=None, to_=None, options=None, force=False):
         remapped.update(results.remapped_ids)
 
 
-    results = UpdateResults(
+    results = ramrod.UpdateResults(
         document=root,
         removed=removed,
         remapped_ids=remapped
