+++ ramrod/cybox/common.py
@@ -1,0 +1,7 @@
+# Copyright (c) 2015, The MITRE Corporation. All rights reserved.
+# See LICENSE.txt for complete terms.
+
+CYBOX_VERSIONS   = ('2.0', '2.0.1', '2.1')
+TAG_CYBOX_MAJOR  = "cybox_major_version"
+TAG_CYBOX_MINOR  = "cybox_minor_version"
+TAG_CYBOX_UPDATE = "cybox_update_version"
