+++ ramrod/stix/common.py
@@ -1,0 +1,4 @@
+# Copyright (c) 2015, The MITRE Corporation. All rights reserved.
+# See LICENSE.txt for complete terms.
+
+STIX_VERSIONS = ('1.0', '1.0.1', '1.1', '1.1.1')
